+++ catapult/__init__.py
@@ -45,7 +45,8 @@ from catapult.conf import ConfigurationStore # noqa
 conf = ConfigurationStore()
 from catapult import i18n # noqa
 from catapult import util # noqa
-from catapult.debug import DebugMixin # noqa
+from catapult.mixins import DebugMixin # noqa
+from catapult.mixins import WindowMixin # noqa
 from catapult.plugin import Plugin # noqa
 from catapult import plugins # noqa
 from catapult.history import History # noqa