+++ NEWS.md
@@ -1,3 +1,8 @@
+PENDING: Catapult 0.1.1
+=======================
+
+* Fix about and preferences dialog modality and positioning
+
 2021-03-16: Catapult 0.1
 ========================